Extract RandomValidNamespace from testutil
